Disable FMA from FFLAGS due to lack of precision in np.dot
The Scipy test scipy/special/tests/test_orthogonal.py fails due to
AssertionError: (Not equal to tolerance). Detailed error in :

https://github.com/scipy/scipy/issues/7535#issuecomment-315498288

The part of the code from test_orthogonal.py where it diverge in
precision is :

    np.dot(v*w, v.T)

Debuging the fail lead us to find out that by removing

    /usr/lib64/haswell/libopenblas_haswellp-r0.2.20.so

The test behaive with proper precision

This patch bisects the haswell falg into: -mavx2 and -mfma. After
experiments find out that -mfma is the flag that creates the lack of
precision

Signed-off-by: Victor Rodriguez <victor.rodriguez.bahena@intel.com>
Tested-by: Gabriel Briones <gbriones@absolute-zero.zpn.intel.com>